requirements.txt
numpy
tifffile
scikit-image
pyyaml
requests
thefuzz